backend/requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==9.2.0
environs[django]==9.3.2
django-phonenumber-field[phonenumbers]==7.0.0
djangorestframework==3.14.0
requests==2.28.1
geopy==2.3.0
rollbar==0.16.3
gunicorn==20.1.0
